Update Domain and DTOs

+++ src/main/java/com/apilivros/Domain/Author.java
@@ -18,7 +18,7 @@ public class Author implements Serializable {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private String nationality;
@@ -30,7 +30,7 @@ public class Author implements Serializable {
     }
     
 
-    public Author(Integer id, String name, String description, String nationality, List<Books> books) {
+    public Author(Long id, String name, String description, String nationality, List<Books> books) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -39,11 +39,11 @@ public class Author implements Serializable {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
