+++ src/main/java/com/apilivros/Dto/ReviewDTO.java
@@ -12,9 +12,9 @@ public class ReviewDTO {
     private String book;
     
     @JsonIgnore
-    private Integer userId;
+    private Long userId;
     @JsonIgnore
-    private Integer bookId;
+    private Long bookId;
 
     public ReviewDTO(){}
     public ReviewDTO(Review review) {
@@ -56,19 +56,19 @@ public class ReviewDTO {
         this.book = book;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getBookId() {
+    public Long getBookId() {
         return bookId;
     }
 
-    public void setBookId(Integer bookId) {
+    public void setBookId(Long bookId) {
         this.bookId = bookId;
     }
 }
